Change every numerical value from Long back to Integer. Add a for loop which checks if the given numbers ever reach the maximum value and if they would surpass it.

+++ src/main/java/com/calculator/Calculator.java
@@ -1,9 +1,9 @@
 package com.calculator;
 
 public class Calculator {
-    private long frst;
-    private long scnd;
-    public Calculator(long x,long y){
+    private int frst;
+    private int scnd;
+    public Calculator(int x,int y){
         this.frst=x;
         this.scnd=y;
     }
